train_code/daily/requirements.txt
joblib==1.3.2
matplotlib==3.8.1
mne==1.5.1
numpy==1.26.1
pydantic==2.4.2
pyedflib==0.1.36
pyserial==3.5
pytest==7.4.3
pyriemann==0.5
scikit_learn==1.3.2
scipy==1.11.3
opencv_python~=4.8.1
pyyaml~=6.0.1
hmmlearn~=0.3.0